pyproject.toml
[project]
name = "voice-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
    "openai-whisper>=20240930",
    "mistralai>=0.0.1",
    "deepgram-sdk>=3.0.0",
    "openai>=2.28.0",
    "sounddevice>=0.5.5",
    "scipy>=1.17.1",
]
